libtool: ignore friend declarations
Friend declarations may declare functions or types which have
de-restricted access to the type.  Such declarations should not be
treated as declarations to decorate (though this will need to change
eventually as the declarations may serve as forward declarations, and
will require to match the actual declaration or definition if it
follows).