libtool: simplify diagnostics emission
We currently only have two diagnostics and do not expect to grow to very
many in any case.  Simply manually expand the code for the cases as we
already need string literals to be passed to the diagnostics.  This
simplifies the callers, and hopefully can eventually consume more of the
building of the diagnostic.