idt: support default visibility attributes in addition to dllimport/dllexport
Address the TODO to handle ELF/MachO visibility attributes similar to the DLLStorage attribute. This extension enables the use of the `idt` tool across a variety of platforms irrespective of the attribute spelling whereas it was previously limited to Windows, PlayStation, and AIX. There is further enhancement still required to support Z/OS which uses `#pragma export` to annotate the ABI surface.

Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>